# docs/swagger.yaml
definitions:
  domain.Post:
    properties:
      AuthorId:
        type: integer
      body:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  domain.PostQuery:
    properties:
      body:
        type: string
      title:
        type: string
    type: object
  domain.SignInInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  domain.SignUpInput:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  domain.Token:
    properties:
      token:
        type: string
    type: object
  domain.UpdatePost:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh User Token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Token'
      summary: Refresh User Token
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: SignIn User
      parameters:
      - description: login user
        in: body
        name: SignIn_input
        required: true
        schema:
          $ref: '#/definitions/domain.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Token'
      summary: SignIn User
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: SignUp User
      parameters:
      - description: new user
        in: body
        name: SignUp_input
        required: true
        schema:
          $ref: '#/definitions/domain.SignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: SignUp User
      tags:
      - Auth
  /post/:
    delete:
      consumes:
      - application/json
      description: Delete a post by ID
      parameters:
      - description: id
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/domain.Post'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "deleted"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a post by ID
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get List of posts
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/domain.Post'
              type: array
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get List of posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create new post with title and content
      parameters:
      - description: new post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/domain.PostQuery'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Post'
      security:
      - ApiKeyAuth: []
      summary: Create new post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update post by ID
      parameters:
      - description: update post
        in: body
        name: updatePost
        required: true
        schema:
          $ref: '#/definitions/domain.UpdatePost'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
      security:
      - ApiKeyAuth: []
      summary: Update post by ID
      tags:
      - posts
  /post/get/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
      security:
      - ApiKeyAuth: []
      summary: Get details of a post
      tags:
      - posts
swagger: "2.0"